middleware/requirements.txt
# /requirements.txt

flask
Authlib
python-dotenv
